fixup: Remove redundant filter, move handlers to separate functions, remove redundant exports, filter members by clientId, remove redundant await
