+++ demo/app/script.ts
@@ -7,6 +7,7 @@ import { getSpaceNameFromUrl } from './utils/url';
 import Spaces from '../../src/Spaces';
 import { renderAvatars, renderSelfAvatar } from './components/avatar-stack';
 import { renderFeatureDisplay } from './components/feature-display';
+import { SpaceMember } from '../../src/Space';
 
 const clientId = nanoid();
 
@@ -14,11 +15,11 @@ const ably = new Ably.Realtime.Promise({ authUrl: `/api/ably-token-request?clien
 
 const spaces = new Spaces(ably);
 
-export const space = spaces.get(getSpaceNameFromUrl(), { offlineTimeout: 60_000 });
+const space = spaces.get(getSpaceNameFromUrl(), { offlineTimeout: 60_000 });
 
-export const selfName = getRandomName();
+const selfName = getRandomName();
 
-const memberIsNotSelf = (member) => member.profileData.name !== selfName;
+const memberIsNotSelf = (member: SpaceMember) => member.clientId !== clientId;
 
 space.on('membersUpdate', (members) => {
   renderAvatars(members.filter(memberIsNotSelf));
@@ -26,7 +27,7 @@ space.on('membersUpdate', (members) => {
 
 renderSelfAvatar(selfName);
 renderFeatureDisplay(space);
-const initialMembers = await space.getMembers();
+const initialMembers = space.getMembers();
 renderAvatars(initialMembers.filter(memberIsNotSelf));
 
 space.enter({ name: selfName });
